backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==2.2.6
pandas==2.3.1
scikit-learn==1.7.1
tensorflow==2.20.0
requests==2.32.4
beautifulsoup4==4.12.2
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
langchain-core>=0.1.7
sentence-transformers==2.2.2
faiss-cpu>=1.7.4
google-generativeai>=0.3.2
python-dotenv==1.0.0
